src/requirements.txt
boto3
mysql-connector-python
python-dotenv